Renaming the StateTransform that is dedicated for MviState
The StateTransform that just mutates the State had the same name as the base StateTransform, this changes the name to not have a name clash.
