--- mvi-declarative/src/main/kotlin/com/adidas/mvi/declarative/transform/SideEffectTransform.kt
@@ -1,16 +1,0 @@
-package com.adidas.mvi.declarative.transform
-
-import com.adidas.mvi.declarative.MviState
-import com.adidas.mvi.sideeffects.SideEffects
-import com.adidas.mvi.transform.StateTransform
-
-public abstract class SideEffectTransform<TState, TSideEffect> :
-    StateTransform<MviState<TState, TSideEffect>> {
-
-    protected abstract fun mutate(sideEffects: SideEffects<TSideEffect>): SideEffects<TSideEffect>
-
-    final override fun reduce(currentState: MviState<TState, TSideEffect>): MviState<TState, TSideEffect> {
-        val sideEffects = currentState.sideEffects
-        return currentState.copy(sideEffects = mutate(sideEffects))
-    }
-}
