+++ mvi-declarative/src/test/kotlin/com/adidas/mvi/declarative/transform/product/FakeProductSideEffectTransform.kt
@@ -1,10 +1,10 @@
 package com.adidas.mvi.declarative.transform.product
 
-import com.adidas.mvi.declarative.transform.SideEffectTransform
+import com.adidas.mvi.declarative.MviState
 import com.adidas.mvi.sideeffects.SideEffects
 
 class FakeProductSideEffectTransform(private val sideEffect: ProductSideEffect) :
-    SideEffectTransform<ProductState, ProductSideEffect>() {
+    MviState.SideEffectTransform<ProductState, ProductSideEffect>() {
 
     override fun mutate(sideEffects: SideEffects<ProductSideEffect>): SideEffects<ProductSideEffect> {
         return sideEffects.add(sideEffect)
