+++ mvi-declarative/src/test/kotlin/com/adidas/mvi/declarative/transform/product/FakeProductStateTransform.kt
@@ -1,8 +1,8 @@
 package com.adidas.mvi.declarative.transform.product
 
-import com.adidas.mvi.declarative.transform.StateTransform
+import com.adidas.mvi.declarative.MviState
 
-class FakeProductStateTransform(private val state: ProductState) : StateTransform<ProductState, ProductSideEffect>() {
+class FakeProductStateTransform(private val state: ProductState) : MviState.StateTransform<ProductState, ProductSideEffect>() {
 
     override fun mutate(currentState: ProductState): ProductState {
         return state
